update projects table and model; change order field type from INTEGER to DOUBLE PRECISION

+++ migrations/20260111000000_create_projects_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS projects (
     end_date_time VARCHAR(10),
     project_status INTEGER NOT NULL,
     version INTEGER,
-    "order" INTEGER,
+    "order" DOUBLE PRECISION,
     creator_id BIGINT NOT NULL,
     updater_id BIGINT,
     create_date_time TIMESTAMP NOT NULL DEFAULT '2022-10-08 00:00:00',
